Add RegressorChain
