+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - `Jaccard` to `metrics`
-- `ClassifierChain` to `multioutput`
+- `ClassifierChain` and `RegressorChain` to `multioutput`
 
 ### Modified
