+++ creme/multioutput/__init__.py
@@ -2,8 +2,10 @@
 Meta-estimators for multi-output classification and regression.
 """
 from .chain import ClassifierChain
+from .chain import RegressorChain
 
 
 __all__ = [
-    'ClassifierChain'
+    'ClassifierChain',
+    'RegressorChain'
 ]
