+++ creme/multioutput/chain.py
@@ -4,7 +4,29 @@ import copy
 from .. import base
 
 
-class ClassifierChain(collections.OrderedDict, base.MultiOutputClassifier):
+__all__ = ['ClassifierChain', 'RegressorChain']
+
+
+class BaseChain(collections.OrderedDict):
+
+    def __init__(self, model, order):
+        super().__init__()
+        for o in order:
+            self[o] = copy.deepcopy(model)
+
+    def fit_one(self, x, y):
+
+        x = copy.copy(x)
+
+        for o, clf in self.items():
+            y_pred = clf.predict_one(x)
+            clf.fit_one(x, y[o])
+            x[o] = y_pred
+
+        return self
+
+
+class ClassifierChain(BaseChain, base.MultiOutputClassifier):
     """A multi-label model that arranges classifiers into a chain.
 
     Example:
@@ -14,7 +36,6 @@ class ClassifierChain(collections.OrderedDict, base.MultiOutputClassifier):
             >>> from creme import feature_selection
             >>> from creme import linear_model
             >>> from creme import metrics
-            >>> from creme import model_selection
             >>> from creme import multioutput
             >>> from creme import preprocessing
             >>> from creme import stream
@@ -29,7 +50,7 @@ class ClassifierChain(collections.OrderedDict, base.MultiOutputClassifier):
             >>> model = feature_selection.VarianceThreshold(threshold=0.01)
             >>> model |= preprocessing.StandardScaler()
             >>> model |= multioutput.ClassifierChain(
-            ...     classifier=linear_model.LogisticRegression(),
+            ...     model=linear_model.LogisticRegression(),
             ...     order=list(range(14))
             ... )
 
@@ -47,22 +68,6 @@ class ClassifierChain(collections.OrderedDict, base.MultiOutputClassifier):
 
     """
 
-    def __init__(self, classifier, order):
-        super().__init__()
-        for o in order:
-            self[o] = copy.deepcopy(classifier)
-
-    def fit_one(self, x, y):
-
-        x = copy.copy(x)
-
-        for o, clf in self.items():
-            y_pred = clf.predict_one(x)
-            clf.fit_one(x, y[o])
-            x[o] = y_pred
-
-        return self
-
     def predict_proba_one(self, x):
 
         x = copy.copy(x)
@@ -71,5 +76,50 @@ class ClassifierChain(collections.OrderedDict, base.MultiOutputClassifier):
         for o, clf in self.items():
             y_pred[o] = clf.predict_proba_one(x)
             x[o] = max(y_pred[o], key=y_pred[o].get)
+
+        return y_pred
+
+
+class RegressorChain(BaseChain, base.MultiOutputRegressor):
+    """A multi-label model that arranges regressor into a chain.
+
+    Example:
+
+        ::
+
+            >>> from creme import linear_model
+            >>> from creme import metrics
+            >>> from creme import model_selection
+            >>> from creme import multioutput
+            >>> from creme import preprocessing
+            >>> from creme import stream
+            >>> from sklearn import datasets
+
+            >>> X_y = stream.iter_sklearn_dataset(
+            ...     dataset=datasets.load_linnerud(),
+            ...     shuffle=True,
+            ...     random_state=42
+            ... )
+
+            >>> model = multioutput.RegressorChain(
+            ...     model=preprocessing.StandardScaler() | linear_model.LinearRegression(),
+            ...     order=[0, 1, 2]
+            ... )
+
+            >>> metric = metrics.RegressionMultiOutput(metrics.MAE())
+
+            >>> model_selection.online_score(X_y, model, metric)
+            MAE: 13.744367
+
+    """
+
+    def predict_one(self, x):
+
+        x = copy.copy(x)
+        y_pred = {}
+
+        for o, clf in self.items():
+            y_pred[o] = clf.predict_one(x)
+            x[o] = y_pred[o]
 
         return y_pred
